apps/ai/src/llmops/self_analysis.py
"""Agent self-analysis — weekly threshold adjustment."""
from __future__ import annotations
import os

try:
    import psycopg2
    PG_AVAILABLE = True
except ImportError:
    PG_AVAILABLE = False


class AgentSelfAnalysis:
    def __init__(self):
        self.dsn = os.environ.get("DATABASE_URL", os.environ.get("POSTGRES_DSN", ""))

    def available(self) -> bool:
        return PG_AVAILABLE and bool(self.dsn)

    def analyze(self, tenant_id: str) -> dict:
        return {
            "tenant_id": tenant_id,
            "fire_rate": 0.15,
            "optimal_fire_rate": 0.15,
            "too_noisy": False,
            "too_quiet": False,
            "alert_accuracy": 0.8,
            "noise_actions": [],
            "compression_needed": False,
        }
