apps/ai/pyproject.toml
[project]
name = "iterateswarm-ai"
version = "0.1.0"
description = "AI Worker for IterateSwarm ChatOps - Temporal Activities"
requires-python = ">=3.11"

dependencies = [
    # Temporal SDK
    "temporalio>=1.11.0",
    # LangChain & LangGraph (AI Agents)
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=1.0.0",
    # OpenAI-compatible client (works with Ollama)
    "openai>=1.0.0",
    # Vector Database
    "qdrant-client>=1.16.0",
    # Caching & Session Store (Sarathi V2.0)
    "redis>=5.0.3",
    # Token Counting
    "tiktoken>=0.7.0",
    # Data Validation
    "pydantic>=2.0.0",
    # HTTP Client
    "httpx>=0.28.0",
    # Structured Logging
    "structlog>=25.0.0",
    # Configuration
    "pyyaml>=6.0.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "dspy-ai>=3.1.3",
    "deepeval>=3.8.4",
    "sentence-transformers>=5.2.3",
    # PostgreSQL async driver
    "asyncpg>=0.29.0",
    # Slack SDK (Saarathi Week 2)
    "slack-sdk>=3.0.0",
    # Bank Statement Parsing
    "pandas>=2.0.0",
    "pdfplumber>=0.10.0",
    "openpyxl>=3.1.0",
    "pytest-timeout>=2.4.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-env>=1.1.3",
    # AI/LLM Testing
    "deepeval>=0.21.0",
    "openai>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["-v"]

[tool.pytest_env]
# Enable running tests without Temporal connection
TEMPORAL_ADDRESS = "localhost:7233"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
